backend/requirements.txt
fastapi
uvicorn[standard]
langchain-openai
langchain-core
langchain_experimental
python-dotenv
pydantic-settings
azure-search-documents>=11.6.0
azure-identity==1.23.1
pypdf==5.9.0
pymupdf==1.26.3
requests>=2.31.0
python-multipart>=0.0.9